+++ jekyll/version-history.markdown
@@ -7,6 +7,7 @@ published: true
 ### v2.0.1
 
  - fix the album covers skewed display when on mobile
+ - fix the tracks displayed under Alex Trebek's page on mobile
  - fix the placement of the LPC logo and link
 
 ---